Fix confirming (#1000)
* Fix confirming

* Fix UI
