+++ lib/screen/interactive_postcard/stamp_preview.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
@@ -5,6 +6,9 @@ import 'package:autonomy_flutter/common/environment.dart';
 import 'package:autonomy_flutter/common/injector.dart';
 import 'package:autonomy_flutter/model/postcard_metadata.dart';
 import 'package:autonomy_flutter/screen/app_router.dart';
+import 'package:autonomy_flutter/screen/detail/artwork_detail_page.dart';
+import 'package:autonomy_flutter/screen/interactive_postcard/postcard_detail_bloc.dart';
+import 'package:autonomy_flutter/screen/interactive_postcard/postcard_detail_state.dart';
 import 'package:autonomy_flutter/screen/interactive_postcard/postcard_view_widget.dart';
 import 'package:autonomy_flutter/service/configuration_service.dart';
 import 'package:autonomy_flutter/service/navigation_service.dart';
@@ -13,12 +17,15 @@ import 'package:autonomy_flutter/util/asset_token_ext.dart';
 import 'package:autonomy_flutter/util/log.dart';
 import 'package:autonomy_flutter/util/postcard_extension.dart';
 import 'package:autonomy_flutter/view/back_appbar.dart';
+import 'package:autonomy_flutter/view/dot_loading_indicator.dart';
 import 'package:autonomy_flutter/view/postcard_button.dart';
 import 'package:autonomy_flutter/view/responsive.dart';
 import 'package:autonomy_theme/autonomy_theme.dart';
+import 'package:autonomy_theme/extensions/theme_extension/moma_sans.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:nft_collection/models/asset_token.dart';
 import 'package:nft_collection/services/tokens_service.dart';
@@ -40,11 +47,13 @@ class _StampPreviewState extends State<StampPreview> {
   Uint8List? postcardData;
   Uint8List? stampedPostcardData;
   int index = 0;
-  bool loading = false;
+  bool confirming = false;
+  Timer? timer;
 
   final _configurationService = injector<ConfigurationService>();
   final _postcardService = injector<PostcardService>();
   final _tokenService = injector<TokensService>();
+  final _navigationService = injector<NavigationService>();
 
   @override
   void initState() {
@@ -54,42 +63,138 @@ class _StampPreviewState extends State<StampPreview> {
   }
 
   @override
+  void dispose() {
+    timer?.cancel();
+    super.dispose();
+  }
+
+  void _refreshPostcard() {
+    log.info("Refresh postcard");
+    context.read<PostcardDetailBloc>().add(PostcardDetailGetInfoEvent(
+          ArtworkIdentity(widget.payload.asset.id, widget.payload.asset.owner),
+        ));
+  }
+
+  void _setTimer() {
+    timer?.cancel();
+    const duration = Duration(seconds: 10);
+    timer = Timer.periodic(duration, (timer) {
+      if (mounted) {
+        _refreshPostcard();
+      }
+    });
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: AppColor.primaryBlack,
-      appBar:
-          getBackAppBar(context, title: "preview_postcard".tr(), onBack: () {
-        Navigator.of(context).pop();
-      }, isWhite: false),
-      body: Padding(
-        padding: ResponsiveLayout.pageHorizontalEdgeInsetsWithSubmitButton,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            PostcardRatio(
-              assetToken: widget.payload.asset,
-              imagePath: widget.payload.imagePath,
-              jsonPath: widget.payload.metadataPath,
-            ),
-            PostcardButton(
-              text: widget.payload.asset.postcardMetadata.isCompleted
-                  ? "complete_postcard_journey_".tr()
-                  : "confirm_your_design".tr(),
-              isProcessing: loading,
-              enabled: !loading,
-              onTap: () async {
-                await _close();
-              },
-            ),
-          ],
+    return WillPopScope(
+      onWillPop: () async {
+        return !confirming;
+      },
+      child: Scaffold(
+        backgroundColor: AppColor.primaryBlack,
+        appBar: getBackAppBar(context,
+            title: "preview_postcard".tr(),
+            onBack: confirming
+                ? null
+                : () {
+                    Navigator.of(context).pop();
+                  },
+            isWhite: false),
+        body: BlocConsumer<PostcardDetailBloc, PostcardDetailState>(
+          listener: (context, state) {
+            if (!(state.isPostcardUpdatingOnBlockchain ||
+                state.isPostcardUpdating)) {
+              if (state.assetToken == null) {
+                return;
+              }
+              _navigationService.popUntilHomeOrSettings();
+              if (!mounted) return;
+              Navigator.of(context).pushNamed(
+                AppRouter.claimedPostcardDetailsPage,
+                arguments:
+                    ArtworkDetailPayload([state.assetToken!.identity], 0),
+              );
+              _configurationService.setAutoShowPostcard(true);
+            }
+          },
+          builder: (context, state) {
+            final assetToken = widget.payload.asset;
+            final imagePath = widget.payload.imagePath;
+            final metadataPath = widget.payload.metadataPath;
+            return Padding(
+              padding:
+                  ResponsiveLayout.pageHorizontalEdgeInsetsWithSubmitButton,
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  PostcardRatio(
+                    assetToken: assetToken,
+                    imagePath: imagePath,
+                    jsonPath: metadataPath,
+                  ),
+                  _postcardAction(state),
+                ],
+              ),
+            );
+          },
         ),
       ),
     );
   }
 
-  Future<void> _close() async {
+  Widget _postcardAction(PostcardDetailState state) {
+    final theme = Theme.of(context);
+    if (!confirming) {
+      return PostcardButton(
+        text: widget.payload.asset.postcardMetadata.isCompleted
+            ? "complete_postcard_journey_".tr()
+            : "confirm_your_design".tr(),
+        onTap: () async {
+          await _onConfirm();
+        },
+      );
+    }
+    if (state.isPostcardUpdatingOnBlockchain) {
+      return PostcardCustomButton(
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.end,
+          children: [
+            Text(
+              "confirming_on_blockchain".tr(),
+              style: theme.textTheme.moMASans700Black14,
+            ),
+            const Padding(
+              padding: EdgeInsets.only(bottom: 5),
+              child: DotsLoading(),
+            ),
+          ],
+        ),
+      );
+    }
+    if (state.isPostcardUpdating) {
+      return PostcardCustomButton(
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.end,
+          children: [
+            Text(
+              "updating_token".tr(),
+              style: theme.textTheme.moMASans700Black14,
+            ),
+            const Padding(
+              padding: EdgeInsets.only(bottom: 5),
+              child: DotsLoading(),
+            ),
+          ],
+        ),
+      );
+    }
+    return const SizedBox();
+  }
+
+  Future<void> _onConfirm() async {
     setState(() {
-      loading = true;
+      confirming = true;
     });
     final imagePath = widget.payload.imagePath;
     final metadataPath = widget.payload.metadataPath;
@@ -106,7 +211,7 @@ class _StampPreviewState extends State<StampPreview> {
     if (walletIndex == null) {
       log.info("[POSTCARD] Wallet index not found");
       setState(() {
-        loading = true;
+        confirming = true;
       });
       return;
     }
@@ -150,14 +255,9 @@ class _StampPreviewState extends State<StampPreview> {
         NftCollectionBloc.eventController.add(
           GetTokensByOwnerEvent(pageKey: PageKey.init()),
         );
+        _setTimer();
       }
-      if (!mounted) return;
-      Navigator.of(context).pushNamed(AppRouter.postcardConfirmingPage,
-          arguments: widget.payload);
     }
-    setState(() {
-      loading = true;
-    });
   }
 }
